[Frontend] Extract shared Realtime codec for RFC 6592 P0a
