+++ vllm_omni/entrypoints/realtime/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Omni project
+
+"""Shared Realtime wire protocol; no model or runtime ownership."""
