+++ vllm_omni/entrypoints/realtime/adapters/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Omni project
+
+"""Execution boundary for consumers of the shared Realtime codec."""
